I improved the script is fixed some bugs
I improved the script is fixed some bugs
